requirements.txt
boto3==1.26.74
botocore==1.29.74
jmespath==1.0.1
p-terraformer==0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
